Added prevention for unwanted user edits

+++ HomeIS/Controllers/ApartmentsController.cs
@@ -23,29 +23,39 @@ namespace HomeIS.Controllers
                 return Redirect("/Account/Login");
             }
 
-            return View(db.Apartments.Where(ap => ap.Owner == db.Users.FirstOrDefault<ApplicationUser>(user => user.UserName == this.User.Identity.Name)));
-            
+            return View(db.Apartments.Where(ap =>
+                ap.Owner == db.Users.FirstOrDefault<ApplicationUser>(user =>
+                    user.UserName == this.User.Identity.Name)));
         }
 
         // GET: Apartments/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (id == null || !Request.IsAuthenticated)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Apartment apartment = db.Apartments.Find(id);
+
+            Apartment apartment = db.Apartments.SingleOrDefault(ap => (ap.ID == id) &&
+                                                                      (ap.Owner.UserName == this.User.Identity.Name));
+
             if (apartment == null)
             {
                 return HttpNotFound();
             }
+
             return View(apartment);
         }
 
         // GET: Apartments/Create
         public ActionResult Create()
         {
-            return View();
+            if (!Request.IsAuthenticated)
+            {
+                return Redirect("/Account/Login");
+            }
+
+            return View(); 
         }
 
         // POST: Apartments/Create
@@ -53,8 +63,18 @@ namespace HomeIS.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,Owner,Location,Description,PropertyValue,PhotoList,Photos,Balcony,Size,FloorNumber,NumberOfRooms")] Apartment apartment)
+        public ActionResult Create([Bind(Include =
+                "ID,Owner,Location,Description,PropertyValue,PhotoList,Photos,Balcony,Size,FloorNumber,NumberOfRooms")]
+            Apartment apartment)
         {
+            if (!Request.IsAuthenticated)
+            {
+                return Redirect("/Account/Login");
+            }
+
+            db.Configuration.LazyLoadingEnabled = false;
+
+
             List<string> PhotoList = new List<string>();
 
             if (ModelState.IsValid && Request.Files.Count > 0)
@@ -70,7 +90,7 @@ namespace HomeIS.Controllers
                 }
 
                 apartment.PhotoList = PhotoList;
-                apartment.Owner = db.Users.SingleOrDefault(s => s.Email == this.User.Identity.Name);
+                apartment.Owner = db.Users.SingleOrDefault(s => s.UserName == this.User.Identity.Name);
 
                 db.Apartments.Add(apartment);
                 db.SaveChanges();
@@ -86,7 +106,8 @@ namespace HomeIS.Controllers
             string absolutePath = AppDomain.CurrentDomain.BaseDirectory + relativePath;
             Directory.CreateDirectory(absolutePath);
 
-            string filename = DateTime.Now.ToFileTime() + (new Random().Next()).ToString() + Path.GetExtension(uploadedFile.FileName);
+            string filename = DateTime.Now.ToFileTime() + (new Random().Next()).ToString() +
+                              Path.GetExtension(uploadedFile.FileName);
             uploadedFile.SaveAs(Path.Combine(absolutePath, filename));
 
             return relativePath + '/' + filename;
@@ -95,15 +116,21 @@ namespace HomeIS.Controllers
         // GET: Apartments/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            db.Configuration.LazyLoadingEnabled = false;
+
+            if (id == null || !Request.IsAuthenticated)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Apartment apartment = db.Apartments.Find(id);
+
+            Apartment apartment = db.Apartments.SingleOrDefault(ap => (ap.ID == id) &&
+                                                                      (ap.Owner.UserName == this.User.Identity.Name));
+
             if (apartment == null)
             {
                 return HttpNotFound();
             }
+
             return View(apartment);
         }
 
@@ -112,7 +139,9 @@ namespace HomeIS.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,Owner,Location,Description,PropertyValue,PhotoList,Photos,Balcony,Size,FloorNumber,NumberOfRooms")] Apartment apartment)
+        public ActionResult Edit([Bind(Include =
+                "ID,Owner,Location,Description,PropertyValue,PhotoList,Photos,Balcony,Size,FloorNumber,NumberOfRooms")]
+            Apartment apartment)
         {
             if (ModelState.IsValid)
             {
@@ -135,27 +164,32 @@ namespace HomeIS.Controllers
                         System.IO.File.Delete(file);
                     }
                 }
-                
+
 
                 db.Entry(apartment).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+
             return View(apartment);
         }
 
         // GET: Apartments/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (id == null || !Request.IsAuthenticated)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Apartment apartment = db.Apartments.Find(id);
+
+            Apartment apartment = db.Apartments.SingleOrDefault(ap => (ap.ID == id) &&
+                                                                      (ap.Owner.UserName == this.User.Identity.Name));
+
             if (apartment == null)
             {
                 return HttpNotFound();
             }
+
             return View(apartment);
         }
 
@@ -164,12 +198,17 @@ namespace HomeIS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Apartment apartment = db.Apartments.Find(id);
+            Apartment apartment = db.Apartments.SingleOrDefault(ap => (ap.ID == id) &&
+                                                                      (ap.Owner.UserName == this.User.Identity.Name));
 
-            apartment.PhotoList.ForEach(photo => System.IO.File.Delete(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, photo)));
+            if (apartment != null)
+            {
+                apartment.PhotoList.ForEach(photo => System.IO.File.Delete(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, photo)));
 
-            db.Apartments.Remove(apartment);
-            db.SaveChanges();
+                db.Apartments.Remove(apartment);
+                db.SaveChanges();
+            }
+
             return RedirectToAction("Index");
         }
 
@@ -179,6 +218,7 @@ namespace HomeIS.Controllers
             {
                 db.Dispose();
             }
+
             base.Dispose(disposing);
         }
 
@@ -191,29 +231,32 @@ namespace HomeIS.Controllers
         public JsonResult ApartmentCountGroupJSON()
         {
             var apartments = from ap in db.Apartments
-                group ap by ap.Location.City into ApartmentGroup
-                select new { City = ApartmentGroup.Key, Count = ApartmentGroup.Count() };
+                group ap by ap.Location.City
+                into ApartmentGroup
+                select new {City = ApartmentGroup.Key, Count = ApartmentGroup.Count()};
 
             return Json(apartments, JsonRequestBehavior.AllowGet);
         }
 
         public JsonResult SizeBalconyMinOrMaxPriceJSON(int Size, bool Balcony, int MinimumPrice)
         {
-            return Json(db.Apartments.Include(t => t.Owner).Where(p => p.Balcony == Balcony && p.Size == Size && 
-                p.PropertyValue >= MinimumPrice).ToList(), JsonRequestBehavior.AllowGet);
+            return Json(db.Apartments.Include(t => t.Owner).Where(p => p.Balcony == Balcony && p.Size == Size &&
+                                                                       p.PropertyValue >= MinimumPrice).ToList(),
+                JsonRequestBehavior.AllowGet);
         }
 
         public JsonResult SizeBalconyPriceRangeJSON(int Size, bool Balcony, int MinimumPrice, int MaximumPrice)
         {
             var QuerySet = db.Apartments.Include(t => t.Owner).Where(p => p.Balcony == Balcony && p.Size == Size &&
-                    p.PropertyValue >= MinimumPrice && p.PropertyValue <= MaximumPrice).ToList();
+                                                                          p.PropertyValue >= MinimumPrice &&
+                                                                          p.PropertyValue <= MaximumPrice).ToList();
             return Json(QuerySet, JsonRequestBehavior.AllowGet);
         }
 
         public JsonResult AmountPerCity(string CityName)
         {
-            var QuerySet = db.Apartments.Where(t => t.Location.City == CityName).GroupBy(p => p.Location.City).
-                Select(g => new { count = g.Count() }).ToList();
+            var QuerySet = db.Apartments.Where(t => t.Location.City == CityName).GroupBy(p => p.Location.City)
+                .Select(g => new {count = g.Count()}).ToList();
 
             return Json(QuerySet, JsonRequestBehavior.AllowGet);
         }
